@@ -6,6 +6,7 @@ use reth_evm_ethereum::RethReceiptBuilder;
 use reth_node_api::{FullNodeComponents, NodeTypes};
 use reth_node_builder::rpc::{EthApiBuilder, EthApiCtx};
 use reth_node_ethereum::EthereumEthApiBuilder;
+use reth_rpc::eth::core::EthRpcConverterFor;
 
 use crate::{
     block::{assembler::TaikoBlockAssembler, factory::TaikoBlockExecutorFactory},
@@ -22,7 +23,7 @@ pub struct TaikoEthApiBuilder(EthereumEthApiBuilder);
 impl TaikoEthApiBuilder {
     /// Creates a new instance of `TaikoEthApiBuilder`.
     pub fn new() -> Self {
-        Self(EthereumEthApiBuilder::default())
+        Self(EthereumEthApiBuilder)
     }
 }
 
@@ -48,25 +49,11 @@ where
         >,
 {
     /// The Ethapi implementation this builder will build.
-    type EthApi = TaikoEthApi<N::Provider, N::Pool, N::Network, N::Evm>;
+    type EthApi = TaikoEthApi<N, EthRpcConverterFor<N>>;
 
     /// Builds the [`TaikoEthApi`] from the given context.
     async fn build_eth_api(self, ctx: EthApiCtx<'_, N>) -> eyre::Result<Self::EthApi> {
-        let api = reth_rpc::EthApiBuilder::new(
-            ctx.components.provider().clone(),
-            ctx.components.pool().clone(),
-            ctx.components.network().clone(),
-            ctx.components.evm_config().clone(),
-        )
-        .eth_cache(ctx.cache)
-        .task_spawner(ctx.components.task_executor().clone())
-        .gas_cap(ctx.config.rpc_gas_cap.into())
-        .max_simulate_blocks(ctx.config.rpc_max_simulate_blocks)
-        .eth_proof_window(ctx.config.eth_proof_window)
-        .fee_history_cache_config(ctx.config.fee_history_cache)
-        .proof_permits(ctx.config.proof_permits)
-        .gas_oracle_config(ctx.config.gas_oracle)
-        .build();
+        let api = ctx.eth_api_builder().build();
 
         Ok(TaikoEthApi(api))
     }
